+++ apps/E2E/src/Badge/pages/BasicBadgePageObject.ts
@@ -1,10 +1,5 @@
-import { BADGE_TESTPAGE, HOMEPAGE_BADGE_BUTTON, BADGE_TEST_COMPONENT, BADGE_SECONDARY_TEST_COMPONENT } from '../consts';
 import { BasePage } from '../../common/BasePage';
-
-export const enum BadgeComponentSelector {
-  PrimaryComponent, //this._primaryComponent
-  SecondaryComponent, //this._secondaryComponent
-}
+import { BADGE_SECONDARY_TEST_COMPONENT, BADGE_TESTPAGE, BADGE_TEST_COMPONENT, HOMEPAGE_BADGE_BUTTON } from '../consts';
 
 class BasicBadgePageObject extends BasePage {
   async getPrimaryComponentAttribute(attribute: string): Promise<string> {