+++ apps/E2E/src/RadioGroupV1/specs/RadioGroupV1.spec.win.ts
@@ -1,18 +1,18 @@
+import { Attribute, BOOT_APP_TIMEOUT, Keys, PAGE_TIMEOUT, RADIOBUTTON_A11Y_ROLE, RADIOGROUP_A11Y_ROLE } from '../../common/consts';
 import NavigateAppPage from '../../common/NavigateAppPage';
-import RadioGroupV1Page, { Radio } from '../pages/RadioGroupV1PageObject';
-import { RADIOBUTTON_A11Y_ROLE, RADIOGROUP_A11Y_ROLE, PAGE_TIMEOUT, BOOT_APP_TIMEOUT, Keys, Attribute } from '../../common/consts';
 import {
+  FIRST_RADIO_ACCESSIBILITY_LABEL,
   RADIOGROUPV1_ACCESSIBILITY_LABEL,
   RADIOGROUPV1_TEST_COMPONENT_LABEL,
-  FIRST_RADIO_ACCESSIBILITY_LABEL,
   SECOND_RADIO_LABEL,
 } from '../consts';
+import RadioGroupV1Page from '../pages/RadioGroupV1PageObject';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('RadioGroupV1/RadioV1 Testing Initialization', function () {
   it('Wait for app load', async () => {
     await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
+    expect(await NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
   });
 
   it('Click and navigate to RadioGroupV1 test page', async () => {
@@ -20,8 +20,8 @@ describe('RadioGroupV1/RadioV1 Testing Initialization', function () {
     await NavigateAppPage.clickAndGoToRadioGroupV1Page();
     await RadioGroupV1Page.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    await expect(await RadioGroupV1Page.isPageLoaded()).toBeTruthy(RadioGroupV1Page.ERRORMESSAGE_PAGELOAD);
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
+    expect(await RadioGroupV1Page.isPageLoaded()).toBeTruthy(RadioGroupV1Page.ERRORMESSAGE_PAGELOAD);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
   });
 });
 
@@ -32,23 +32,23 @@ describe('RadioGroupV1/RadioV1 Accessibility Testing', () => {
   });
 
   it('Validate RadioGroup\'s "accessibilityRole" defaults to "ControlType.List".', async () => {
-    await expect(
+    expect(
       await RadioGroupV1Page.compareAttribute(RadioGroupV1Page._primaryComponent, Attribute.AccessibilityRole, RADIOGROUP_A11Y_ROLE),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Validate Radio\'s "accessibilityRole" defaults to "ControlType.RadioButton".', async () => {
-    await expect(
-      await RadioGroupV1Page.compareAttribute(RadioGroupV1Page.getRadio(Radio.First), Attribute.AccessibilityRole, RADIOBUTTON_A11Y_ROLE),
+    expect(
+      await RadioGroupV1Page.compareAttribute(RadioGroupV1Page.getRadio('First'), Attribute.AccessibilityRole, RADIOBUTTON_A11Y_ROLE),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Set RadioGroup "accessibilityLabel" prop. Validate "accessibilityLabel" value propagates to "Name" element attribute.', async () => {
-    await expect(
+    expect(
       await RadioGroupV1Page.compareAttribute(
         RadioGroupV1Page._primaryComponent,
         Attribute.AccessibilityLabel,
@@ -56,11 +56,11 @@ describe('RadioGroupV1/RadioV1 Accessibility Testing', () => {
       ),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Do not set RadioGroup "accessibilityLabel" prop. Validate "Name" element attribute defaults to current RadioGroup label.', async () => {
-    await expect(
+    expect(
       await RadioGroupV1Page.compareAttribute(
         RadioGroupV1Page._secondaryComponent,
         Attribute.AccessibilityLabel,
@@ -68,27 +68,27 @@ describe('RadioGroupV1/RadioV1 Accessibility Testing', () => {
       ),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Set Radio "accessibilityLabel" prop. Validate "accessibilityLabel" value propagates to "Name" element attribute.', async () => {
-    await expect(
+    expect(
       await RadioGroupV1Page.compareAttribute(
-        RadioGroupV1Page.getRadio(Radio.First),
+        RadioGroupV1Page.getRadio('First'),
         Attribute.AccessibilityLabel,
         FIRST_RADIO_ACCESSIBILITY_LABEL,
       ),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Do not set Radio "accessibilityLabel" prop. Validate "Name" element attribute defaults to current Radio label.', async () => {
-    await expect(
-      await RadioGroupV1Page.compareAttribute(RadioGroupV1Page.getRadio(Radio.Second), Attribute.AccessibilityLabel, SECOND_RADIO_LABEL),
+    expect(
+      await RadioGroupV1Page.compareAttribute(RadioGroupV1Page.getRadio('Second'), Attribute.AccessibilityLabel, SECOND_RADIO_LABEL),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 });
 
@@ -102,72 +102,69 @@ describe('RadioGroupV1 Functional Testing', async () => {
 
   it('Click on a Radio. Validate that it changes state from unselected to selected.', async () => {
     /* Validate the Radio is not initially selected */
-    await expect(await RadioGroupV1Page.isRadioSelected(Radio.Second)).toBeFalsy(
+    expect(await RadioGroupV1Page.isRadioSelected('Second')).toBeFalsy(
       'Expected radio #2 to be unselected at test start, but #2 was initially selected.',
     );
 
     /* Click on the Radio to select it */
-    await RadioGroupV1Page.click(RadioGroupV1Page.getRadio(Radio.Second));
+    await RadioGroupV1Page.click(RadioGroupV1Page.getRadio('Second'));
 
     /* Validate the Radio is selected */
-    await expect(
-      await RadioGroupV1Page.waitForRadioSelected(Radio.Second, 'Expected radio #2 to be selected by click, but #2 remained unselected.'),
+    expect(
+      await RadioGroupV1Page.waitForRadioSelected('Second', 'Expected radio #2 to be selected by click, but #2 remained unselected.'),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Navigate to unselected radio using "DOWN ARROW" key. Validate state changes from unselected to selected.', async () => {
     // Presses the ArrowDown key while the first (A) Radio is selected
-    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio(Radio.First), [Keys.ARROW_DOWN]);
+    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio('First'), [Keys.ARROW_DOWN]);
 
     /* Validate the Radio is selected */
-    await expect(
-      await RadioGroupV1Page.waitForRadioSelected(Radio.Second, 'Expected radio #2 to be selected by a "DOWN ARROW" input from radio #1.'),
+    expect(
+      await RadioGroupV1Page.waitForRadioSelected('Second', 'Expected radio #2 to be selected by a "DOWN ARROW" input from radio #1.'),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Navigate to unselected radio using "DOWN ARROW" key. Validate disabled Radio is skipped.', async () => {
     // Presses the ArrowDown key while the second (B) Radio is selected
-    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio(Radio.Second), [Keys.ARROW_DOWN]);
+    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio('Second'), [Keys.ARROW_DOWN]);
 
     /* Validate the Radio is selected */
-    await expect(
+    expect(
       await RadioGroupV1Page.waitForRadioSelected(
-        Radio.Fourth,
+        'Fourth',
         'Expected radio #4 to be selected by a "DOWN ARROW" input from radio #2 and radio #3 (disabled) to be skipped.',
       ),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Press "DOWN ARROW" on the last Radio of a RadioGroup. Validate circular navigation functions correctly.', async () => {
     // Presses the ArrowDown key while the fourth (D) Radio is selected
-    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio(Radio.Fourth), [Keys.ARROW_DOWN]);
+    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio('Fourth'), [Keys.ARROW_DOWN]);
 
     /* Validate the Radio is selected */
-    await expect(
-      await RadioGroupV1Page.waitForRadioSelected(Radio.First, 'Expected radio #1 to be selected by a "DOWN ARROW" input from radio #4.'),
+    expect(
+      await RadioGroupV1Page.waitForRadioSelected('First', 'Expected radio #1 to be selected by a "DOWN ARROW" input from radio #4.'),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 
   it('Press "TAB" on a Radio. Validate Radio in the next RadioGroup is focused.', async () => {
     // Presses the Tab key while the second (B) Radio is selected in first RadioGroup
-    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio(Radio.Second), [Keys.TAB]);
+    await RadioGroupV1Page.sendKeys(RadioGroupV1Page.getRadio('Second'), [Keys.TAB]);
 
     /* Validate the Radio is not focused */
-    await expect(
-      await RadioGroupV1Page.waitForRadioFocused(
-        Radio.Fifth,
-        'Expected radio #5 to be focused by a "TAB" input from previous radio group.',
-      ),
+    expect(
+      await RadioGroupV1Page.waitForRadioFocused('Fifth', 'Expected radio #5 to be focused by a "TAB" input from previous radio group.'),
     ).toBeTruthy();
 
-    await expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
+    expect(await RadioGroupV1Page.didAssertPopup()).toBeFalsy(RadioGroupV1Page.ERRORMESSAGE_ASSERT);
   });
 });