+++ packages/components/Badge/src/Badge.types.ts
@@ -1,10 +1,10 @@
-import { ColorValue, FlexStyle } from 'react-native';
-import { TextProps } from '@fluentui-react-native/text';
-import { FontTokens, IBorderTokens, IColorTokens, IShadowTokens, LayoutTokens } from '@fluentui-react-native/tokens';
-import { IViewProps } from '@fluentui-react-native/adapters';
-import { IconProps, IconSourcesType } from '@fluentui-react-native/icon';
-import { ShadowProps } from '@fluentui-react-native/experimental-shadow';
-import { ShadowToken } from '@fluentui-react-native/theme-types';
+import type { ColorValue, FlexStyle } from 'react-native';
+import type { TextProps } from '@fluentui-react-native/text';
+import type { FontTokens, IBorderTokens, IColorTokens, IShadowTokens, LayoutTokens } from '@fluentui-react-native/tokens';
+import type { IViewProps } from '@fluentui-react-native/adapters';
+import type { IconProps, IconSourcesType } from '@fluentui-react-native/icon';
+import type { ShadowProps } from '@fluentui-react-native/experimental-shadow';
+import type { ShadowToken } from '@fluentui-react-native/theme-types';
 
 export const badgeName = 'Badge';
 export const BadgeSizes = ['tiny', 'extraSmall', 'small', 'medium', 'large', 'extraLarge'] as const;