+++ packages/components/Button/src/FAB/FAB.styling.ts
@@ -1,5 +1,7 @@
-import { fabName, FABProps, FABSlotProps, FABTokens } from './FAB.types';
-import { Theme, UseStylingOptions, buildProps } from '@fluentui-react-native/framework';
+import type { FABProps, FABSlotProps, FABTokens } from './FAB.types';
+import { fabName } from './FAB.types';
+import type { Theme, UseStylingOptions } from '@fluentui-react-native/framework';
+import { buildProps } from '@fluentui-react-native/framework';
 import { borderStyles, layoutStyles, fontStyles, shadowStyles } from '@fluentui-react-native/tokens';
 import { getTextMarginAdjustment } from '@fluentui-react-native/styling-utils';
 import { Platform } from 'react-native';