+++ packages/components/Checkbox/src/shouldUseThickCheckmark.ts
@@ -1,4 +1,4 @@
-import { Theme } from '@fluentui-react-native/framework';
+import type { Theme } from '@fluentui-react-native/framework';
 
 export function shouldUseThickCheckmark(_theme: Theme): boolean {
   return false;