+++ packages/components/PersonaCoin/src/PersonaCoin.settings.ts
@@ -1,5 +1,6 @@
-import { IComposeSettings } from '@uifabricshared/foundation-compose';
-import { personaCoinName, IPersonaCoinType } from './PersonaCoin.types';
+import type { IComposeSettings } from '@uifabricshared/foundation-compose';
+import type { IPersonaCoinType } from './PersonaCoin.types';
+import { personaCoinName } from './PersonaCoin.types';
 import { convertCoinColorFluent } from './PersonaCoin.helpers';
 
 export const settings: IComposeSettings<IPersonaCoinType> = [