+++ packages/deprecated/theming-react-native/src/Global.ts
@@ -1,6 +1,6 @@
-import { IPartialTheme, ITheme, ThemeRegistry } from './Theme.types';
+import type { IPartialTheme, ITheme, ThemeRegistry } from './Theme.types';
 import { createPlatformThemeRegistry } from './platform';
-import { IProcessTheme, IThemeEventListener } from '@uifabricshared/theme-registry';
+import type { IProcessTheme, IThemeEventListener } from '@uifabricshared/theme-registry';
 
 let _registry: ThemeRegistry;