+++ packages/experimental/ActivityIndicator/src/index.ts
@@ -1,1 +1,1 @@
-export * from './ActivityIndicator';
+export { ActivityIndicator } from './ActivityIndicator';