+++ packages/experimental/RadioGroup/src/RadioGroup/RadioGroup.types.ts
@@ -1,8 +1,8 @@
 import type { IViewProps } from '@fluentui-react-native/adapters';
-import { TextProps } from '@fluentui-react-native/text';
-import { IForegroundColorTokens, FontTokens } from '@fluentui-react-native/tokens';
-import { FocusZoneProps } from '@fluentui-react-native/focus-zone';
-import { ColorValue, ViewStyle } from 'react-native';
+import type { TextProps } from '@fluentui-react-native/text';
+import type { IForegroundColorTokens, FontTokens } from '@fluentui-react-native/tokens';
+import type { FocusZoneProps } from '@fluentui-react-native/focus-zone';
+import type { ColorValue, ViewStyle } from 'react-native';
 
 export const radioGroupName = 'RadioGroup';