+++ packages/framework/use-slot/src/index.ts
@@ -1,4 +1,6 @@
-export * from './renderSlot';
-export * from './stagedComponent';
-export * from './useSlot';
-export * from './withSlots';
+export { renderSlot } from './renderSlot';
+export type { NativeReactType, SlotFn } from './renderSlot';
+export { stagedComponent } from './stagedComponent';
+export type { ComposableFunction, FinalRender, StagedRender } from './stagedComponent';
+export { useSlot } from './useSlot';
+export { withSlots } from './withSlots';