+++ packages/theming/win32-theme/src/createOfficeAliasTokens.ts
@@ -1,5 +1,5 @@
 import { getOfficeAliasTokens, getOfficeShadowTokens } from './getOfficeTokens';
-import { AliasColorTokens, ThemeShadowDefinition } from '@fluentui-react-native/theme-types';
+import type { AliasColorTokens, ThemeShadowDefinition } from '@fluentui-react-native/theme-types';
 import { memoize } from '@fluentui-react-native/memo-cache';
 import { mapPipelineToShadow, mapPipelineToTheme } from '@fluentui-react-native/theming-utils';