+++ packages/experimental/Drawer/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @fluentui-react-native/experimental-drawer
 
-This log was last generated on Mon, 30 Jan 2023 07:12:38 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Feb 2023 01:29:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.41
+
+Thu, 02 Feb 2023 01:29:38 GMT
+
+### Patches
+
+- Add eslint to packages missing a config (30809111+acoates-ms@users.noreply.github.com)
+- Bump @fluentui-react-native/framework to v0.8.39
+
+## 0.2.40
+
+Wed, 01 Feb 2023 22:33:17 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/component-cache to v1.4.3
+- Bump @fluentui-react-native/framework to v0.8.38
 
 ## 0.2.39